requested change - directory for models + updated header file
