+++ nn_from_scratch/hardware/model/model.h
@@ -1,9 +1,12 @@
-#include <cstdint>
+#include <stdint.h>
 
-#define INPUT_SIZE {input_size}
-#define N_LAYERS {n_layers}
+#define INPUT_SIZE 1
+#define N_LAYERS 3
 
-{layers_size}
+#define LAYER_0_SIZE 4
+#define LAYER_1_SIZE 2
+#define LAYER_2_SIZE 1
+
 
 enum ActivationType {
     LINEAR,
